Refactored `user` package to `model` for consistency and updated package imports accordingly. Upgraded Spring Boot dependency to version 3.3.4. Added system scope configuration for required libraries.
