+++ TourGuide/src/main/java/com/openclassrooms/tourguide/model/User.java
@@ -1,4 +1,4 @@
-package com.openclassrooms.tourguide.user;
+package com.openclassrooms.tourguide.model;
 
 import java.util.ArrayList;
 import java.util.Date;
