+++ TourGuide/src/main/java/com/openclassrooms/tourguide/module/TourGuideModule.java
@@ -1,4 +1,4 @@
-package com.openclassrooms.tourguide;
+package com.openclassrooms.tourguide.module;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
